requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
spotipy==2.23.0
python-dotenv==1.0.1